chore: change ci steps (#531)
