+++ .nvmrc
@@ -1,1 +1,1 @@
-18
+20.10.0
